+++ src/comps/Features/Darkmode.svelte
@@ -77,53 +77,61 @@
   });
 </script>
 
-<svg
-  class="darkmode {isReady ? 'is-ready' : ''}"
-  xmlns="http://www.w3.org/2000/svg"
-  width="24"
-  height="24"
-  viewBox="0 0 24 24"
-  fill="none"
-  stroke-width="2"
-  stroke-linecap="round"
-  stroke-linejoin="round"
-  stroke="currentColor"
-  on:click={toggleDarkMode}
-  style="transform:rotate({$svgContainerProps.transform}deg)"
->
-  <mask id="myMask2">
-    <rect x="0" y="0" width="100%" height="100%" fill="white" />
-    <circle cx={$maskedCircleProps.cx} cy={$maskedCircleProps.cy} r="9" fill="black" />
-  </mask>
-  <circle cx="12" cy="12" fill="white" mask="url(#myMask2)" r={$centerCircleProps.r > 0 ? $centerCircleProps.r : 0} />
-  <g stroke="currentColor" style="opacity:{$linesProps.opacity || 0}">
-    <line x1="12" y1="1" x2="12" y2="3" />
-    <line x1="12" y1="21" x2="12" y2="23" />
-    <line x1="4.22" y1="4.22" x2="5.64" y2="5.64" />
-    <line x1="18.36" y1="18.36" x2="19.78" y2="19.78" />
-    <line x1="1" y1="12" x2="3" y2="12" />
-    <line x1="21" y1="12" x2="23" y2="12" />
-    <line x1="4.22" y1="19.78" x2="5.64" y2="18.36" />
-    <line x1="18.36" y1="5.64" x2="19.78" y2="4.22" />
-  </g>
-</svg>
+<button class="circle extra {isReady ? 'is-ready' : ''}" on:click={toggleDarkMode}>
+  <svg
+    class="darkmode"
+    xmlns="http://www.w3.org/2000/svg"
+    width="24"
+    height="24"
+    viewBox="0 0 24 24"
+    fill="none"
+    stroke-width="2"
+    stroke-linecap="round"
+    stroke-linejoin="round"
+    stroke="currentColor"
+    style="transform:rotate({$svgContainerProps.transform}deg)"
+  >
+    <mask id="myMask2">
+      <rect x="0" y="0" width="100%" height="100%" fill="white" />
+      <circle cx={$maskedCircleProps.cx} cy={$maskedCircleProps.cy} r="9" fill="black" />
+    </mask>
+    <circle
+      cx="12"
+      cy="12"
+      fill="currentColor"
+      mask="url(#myMask2)"
+      r={$centerCircleProps.r > 0 ? $centerCircleProps.r : 0}
+    />
+    <g stroke="currentColor" style="opacity:{$linesProps.opacity || 0}">
+      <line x1="12" y1="1" x2="12" y2="3" />
+      <line x1="12" y1="21" x2="12" y2="23" />
+      <line x1="4.22" y1="4.22" x2="5.64" y2="5.64" />
+      <line x1="18.36" y1="18.36" x2="19.78" y2="19.78" />
+      <line x1="1" y1="12" x2="3" y2="12" />
+      <line x1="21" y1="12" x2="23" y2="12" />
+      <line x1="4.22" y1="19.78" x2="5.64" y2="18.36" />
+      <line x1="18.36" y1="5.64" x2="19.78" y2="4.22" />
+    </g>
+  </svg>
+</button>
 
 <style lang="scss">
-  @import '../../scss/variables';
-
-  .darkmode {
+  button {
     transition: opacity 1000ms ease-in-out;
     will-change: opacity;
+    opacity: 0;
     position: fixed;
     z-index: 1337;
-    right: 16rem;
+    right: 84rem;
     bottom: 16rem;
-    user-select: none;
-    opacity: 0;
-    cursor: pointer;
 
     &.is-ready {
       opacity: 1 !important;
     }
+  }
+
+  .darkmode {
+    user-select: none;
+    cursor: pointer;
   }
 </style>