+++ src/stores/api/history-data.ts
@@ -1,6 +1,7 @@
 import { storedProducts } from '../products';
 import { sortedStates, storedActiveSelection } from '../states';
-import { ID_STATE_ANNOUNCEMENT, ID_STATE_AVAILABLE, ID_STATE_COMING_SOON, LOADED, LOADING } from '../../_interfaces';
+import { LOADED, LOADING } from '../../_interfaces';
+import { setFlags } from './product-data';
 
 export const loadHistoryData = async () => {
   storedActiveSelection.update(store => {
@@ -30,19 +31,7 @@ export const loadHistoryData = async () => {
       const state = sortedStates.find(state => state.id === product.state);
       product.state = state;
 
-      // flags
-      const historyStates = Object.values(product.history);
-      const lastHistory = historyStates[historyStates.length - 1];
-      const beforeLastHistory = historyStates[historyStates.length - 2];
-      const beforeBeforeLastHistory = historyStates[historyStates.length - 3];
-
-      product.isNewSoon = lastHistory === ID_STATE_COMING_SOON && beforeLastHistory === ID_STATE_ANNOUNCEMENT;
-      product.isNew =
-        (lastHistory === ID_STATE_AVAILABLE && beforeLastHistory === ID_STATE_ANNOUNCEMENT) ||
-        (lastHistory === ID_STATE_AVAILABLE &&
-          beforeLastHistory === ID_STATE_COMING_SOON &&
-          beforeBeforeLastHistory === ID_STATE_ANNOUNCEMENT);
-      product.isHot = historyStates.filter(state => state === ID_STATE_AVAILABLE).length >= 3;
+      setFlags(product);
 
       // new products with history is ready
       return product;